requirements.txt
discord.py>=2.3.0
toml>=0.10.0 